Remove depenendcy on sc-consensus-slots
